refactor non root login switch by new design
