+++ Makefile
@@ -46,4 +46,4 @@ endif
 	go build -o bin/tidb-dashboard -ldflags '$(LDFLAGS)' -tags "${BUILD_TAGS}" cmd/tidb-dashboard/main.go
 
 run:
-	bin/tidb-dashboard --debug --experimental --non-root-login
+	bin/tidb-dashboard --debug --experimental
