+++ pkg/apiserver/info/info.go
@@ -23,6 +23,7 @@ import (
 	"go.uber.org/fx"
 
 	"github.com/pingcap/tidb-dashboard/pkg/apiserver/conprof"
+	"github.com/pingcap/tidb-dashboard/pkg/apiserver/nonrootlogin"
 	"github.com/pingcap/tidb-dashboard/pkg/apiserver/user"
 	"github.com/pingcap/tidb-dashboard/pkg/apiserver/utils"
 	"github.com/pingcap/tidb-dashboard/pkg/config"
@@ -62,6 +63,7 @@ type InfoResponse struct { //nolint
 	EnableTelemetry    bool          `json:"enable_telemetry"`
 	EnableExperimental bool          `json:"enable_experimental"`
 	EnableConprof      bool          `json:"enable_conprof"`
+	EnableNonRootLogin bool          `json:"enable_non_root_login"`
 }
 
 // @ID infoGet
@@ -76,6 +78,7 @@ func (s *Service) infoHandler(c *gin.Context) {
 		EnableTelemetry:    s.params.Config.EnableTelemetry,
 		EnableExperimental: s.params.Config.EnableExperimental,
 		EnableConprof:      conprof.IsFeatureEnable(s.params.Config),
+		EnableNonRootLogin: nonrootlogin.IsFeatureEnable(s.params.Config),
 	}
 	c.JSON(http.StatusOK, resp)
 }
