+++ pkg/apiserver/nonrootlogin/feature_enable.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package conprof
+package nonrootlogin
 
 import (
 	"github.com/pingcap/tidb-dashboard/pkg/apiserver/utils"
