+++ pkg/apiserver/user/auth.go
@@ -27,11 +27,9 @@ import (
 	"github.com/gtank/cryptopasta"
 	"github.com/joomcode/errorx"
 	"github.com/pingcap/log"
-	"go.uber.org/fx"
 	"go.uber.org/zap"
 
 	"github.com/pingcap/tidb-dashboard/pkg/apiserver/utils"
-	"github.com/pingcap/tidb-dashboard/pkg/config"
 )
 
 var (
@@ -41,13 +39,7 @@ var (
 	ErrSignInOther         = ErrNSSignIn.NewType("other")
 )
 
-type ServiceParams struct {
-	fx.In
-	Config *config.Config
-}
-
 type AuthService struct {
-	params         ServiceParams
 	middleware     *jwt.GinJWTMiddleware
 	authenticators map[utils.AuthType]Authenticator
 }
@@ -89,7 +81,7 @@ func (a BaseAuthenticator) SignOutInfo(u *utils.SessionUser, redirectURL string)
 	return &SignOutInfo{}, nil
 }
 
-func NewAuthService(p ServiceParams) *AuthService {
+func NewAuthService() *AuthService {
 	var secret *[32]byte
 
 	secretStr := os.Getenv("DASHBOARD_SESSION_SECRET")
@@ -106,7 +98,6 @@ func NewAuthService(p ServiceParams) *AuthService {
 	}
 
 	service := &AuthService{
-		params:         p,
 		middleware:     nil,
 		authenticators: map[utils.AuthType]Authenticator{},
 	}
@@ -295,7 +286,6 @@ func (s *AuthService) RegisterAuthenticator(typeID utils.AuthType, a Authenticat
 
 type GetLoginInfoResponse struct {
 	SupportedAuthTypes []int `json:"supported_auth_types"`
-	EnableNonRootLogin bool  `json:"enable_non_root_login"`
 }
 
 // @ID userGetLoginInfo
@@ -317,7 +307,6 @@ func (s *AuthService) getLoginInfoHandler(c *gin.Context) {
 	sort.Ints(supportedAuth)
 	resp := GetLoginInfoResponse{
 		SupportedAuthTypes: supportedAuth,
-		EnableNonRootLogin: s.params.Config.EnableNonRootLogin,
 	}
 	c.JSON(http.StatusOK, resp)
 }
